add some change related cart

+++ app/controllers/api/v1/wishlist_items_controller.rb
@@ -20,16 +20,33 @@ class Api::V1::WishlistItemsController < ApplicationController
     head :no_content
   end
 
+  # def add_to_cart
+  #   wishlist_item = @wishlist.wishlist_items.find_by(product_item_id: params[:product_item_id])
+  #   if wishlist_item
+  #     cart = Cart.find_or_create_by(user: @wishlist.user)
+  #     cart_item = cart.cart_items.find_or_initialize_by(product_item: wishlist_item.product_item)
+  #     cart_item.quantity = cart_item.new_record? ? 1 : cart_item.quantity
+  #     # cart_item.quantity = (cart_item.quantity || 0) + 1
+  #     cart_item.save
+  #     wishlist_item.destroy
+  #     render json: cart, include: { cart_items: { include: :product_item } }
+  #   else
+  #     render json: { error: 'Item not found in wishlist' }, status: :not_found
+  #   end
+  # end
+
   def add_to_cart
-    wishlist_item = @wishlist.wishlist_items.find_by(product_item_id: params[:product_item_id])
+    wishlist_item = @wishlist.wishlist_items.find_by(product_item_variant_id: params[:product_item_variant_id])
     if wishlist_item
       cart = Cart.find_or_create_by(user: @wishlist.user)
-      cart_item = cart.cart_items.find_or_initialize_by(product_item: wishlist_item.product_item)
-      cart_item.quantity = cart_item.new_record? ? 1 : cart_item.quantity
-      # cart_item.quantity = (cart_item.quantity || 0) + 1
-      cart_item.save
-      wishlist_item.destroy
-      render json: cart, include: { cart_items: { include: :product_item } }
+      cart_item = cart.cart_items.find_or_initialize_by(product_item_variant: wishlist_item.product_item_variant)
+      cart_item.quantity = cart_item.new_record? ? 1 : cart_item.quantity + 1
+      if cart_item.save
+        wishlist_item.destroy
+        render json: cart, include: { cart_items: { include: { product_item_variant: { include: :product_item } } } }
+      else
+        render json: cart_item.errors, status: :unprocessable_entity
+      end
     else
       render json: { error: 'Item not found in wishlist' }, status: :not_found
     end
